requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
lxml>=4.9.0
pandas>=2.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
click>=8.1.0
pydantic>=2.4.0
aiohttp>=3.8.0
playwright>=1.40.0
fake-useragent>=1.4.0
